Cleaned up main.c from other util files, need to better implement the storage of texture images
